+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 - Run the GUI in background.
 - Abbreviation for Erlang.
+- Option to run the GUI in foreground
 
 
 ## 0.1.1 - 2018-02-23